// src/modules/documents/models/folders.enum.ts
export enum Folders {
  PM_INSTRUCTIONS = 'PM Instructions',
  COMMERCIAL_DOCS = 'Commercial Docs',
  PM_REPORTS = 'PM Reports',
  IMAGES = 'Images',
  VIDEOS = 'Videos',
  DOCUMENT_UPLOAD = 'Document Upload',
  ACTIVITY = 'Activity',
  PROCEDURES = 'Procedures',
  SUBMIT_FOR_REVIEW = 'Submit For Review',
  TEST_REPORTS = 'Test Reports',
  SERVICE_REPORTS = 'Service Reports',
  PRIVATE = 'Private',
  OTHERS = 'Others',
  INSTRUCTION_MANUAL = 'Instruction Manual',
  FINAL_DRAWINGS = 'Final Drawings',
  ELECTRICAL_PACKAGES = 'Electrical Packages',
  ADDENDUMS = 'Addendums',
  POWER_STUDIES = 'Power Studies',
}